+++ src/main/java/com/example/moneytrack/dto/DepositRequest.java
@@ -3,11 +3,9 @@ package com.example.moneytrack.dto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Getter
 @NoArgsConstructor
 public class DepositRequest {
     private String accountNumber;
-    private BigDecimal amount;
+    private Long amount;
 }
